+++ tests/testthat/test_codebook.R
@@ -269,10 +269,10 @@ test_that("codebook table generation", {
   bfi$age <- 1:nrow(bfi)
   expect_silent(cb <- codebook_table(bfi))
   expect_identical(names(cb),
-    c("name", "label", "type", "type_options", "data_type",
+    c("name", "label", "type", "type_options", "data_type", "ordered",
       "value_labels", "optional", "scale_item_names", "item_order",
       "missing", "complete", "n",  "empty", "n_unique",
-      "count", "median", "min", "max", "mean", "sd",
+      "top_counts", "count", "median", "min", "max", "mean", "sd",
       "p0", "p25", "p50", "p75", "p100", "hist"))
   expect_equal(nrow(cb), ncol(bfi))
   expect_identical(cb$name, names(bfi))